BLD: Include python debug build for Travis-ci
Our current CI setup does not run with a debug version of Python for
Travis-CI. This has led to several issues that may be related to
Python's debug symbols [1, 2]. This commit adds a build using with
python-dbg to the Travis-CI matrix.

[1] https://github.com/scipy/scipy/issues/9245
[2] https://github.com/scipy/scipy/issues/10388

The debug build requires python3-bdg and python3-dev in addition to the
usual SciPy compilers/tools etc. To include this we define an anchor for
the common packages and reference it in the debug build's matrix entry.
This leads to an error as the packages are not yet defined.

To run the debug build we must explicitly construct a venv, this was not
previously required as Travis-CI constructs a venv for the specified
python. However as the debug version cannot be specified a new virtual
environment must now be manually constructed.